Create Markowitz_allocator_wr.py
Creation of the Markowitz_Allocator_WR class in the case of an allocation problem with only one non-risky asset (and the rest of the assets are all risky). This class receives the historical observations in the form of a matrix, the rate of the non-risky assets and may (or may not) also receive the names of each asset to identify their respective weights.